move classes to subpackages for better separation
