+++ gazeplay/src/main/java/net/gazeplay/GazePlay.java
@@ -15,6 +15,13 @@ import net.gazeplay.commons.ui.Translator;
 import net.gazeplay.commons.utils.games.BackgroundMusicManager;
 import net.gazeplay.commons.utils.multilinguism.Multilinguism;
 import net.gazeplay.components.CssUtil;
+import net.gazeplay.ui.scenes.configuration.ConfigurationContext;
+import net.gazeplay.ui.scenes.gamemenu.HomeMenuScreen;
+import net.gazeplay.ui.scenes.ingame.GameContext;
+import net.gazeplay.ui.scenes.stats.AreaOfInterest;
+import net.gazeplay.ui.scenes.stats.ScanpathView;
+import net.gazeplay.ui.scenes.stats.StatsContext;
+import net.gazeplay.ui.scenes.userselect.UserProfilContext;
 import org.springframework.stereotype.Component;
 
 @Slf4j
