+++ gazeplay/src/main/java/net/gazeplay/GazePlayFxApp.java
@@ -15,7 +15,9 @@ import javafx.stage.WindowEvent;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.commons.configuration.ActiveConfigurationContext;
 import net.gazeplay.commons.configuration.Configuration;
+import net.gazeplay.commons.utils.ApplicationIconImageLocator;
 import net.gazeplay.components.CssUtil;
+import net.gazeplay.latestnews.LatestNewPopup;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
