+++ gazeplay/src/main/java/net/gazeplay/commons/app/GazePlayAnimatedLogo.java
@@ -1,4 +1,4 @@
-package net.gazeplay;
+package net.gazeplay.commons.app;
 
 import javafx.animation.*;
 import javafx.geometry.Pos;
