+++ gazeplay/src/main/java/net/gazeplay/commons/utils/stats/StatsDisplay.java
@@ -15,7 +15,7 @@ import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.GazePlay;
-import net.gazeplay.StatsContext;
+import net.gazeplay.ui.scenes.stats.StatsContext;
 import net.gazeplay.commons.utils.FixationPoint;
 import net.gazeplay.commons.utils.HomeButton;
 import net.gazeplay.stats.ShootGamesStats;
