+++ gazeplay/src/main/java/net/gazeplay/gameslocator/GamesLocator.java
@@ -1,5 +1,6 @@
-package net.gazeplay;
+package net.gazeplay.gameslocator;
 
+import net.gazeplay.GameSpec;
 import net.gazeplay.commons.ui.Translator;
 
 import java.util.List;
