+++ gazeplay/src/main/java/net/gazeplay/ui/scenes/gamemenu/GameButtonOrientation.java
@@ -1,4 +1,4 @@
-package net.gazeplay;
+package net.gazeplay.ui.scenes.gamemenu;
 
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.commons.configuration.Configuration;
