+++ gazeplay/src/test/java/net/gazeplay/ApplicationIconImageLocatorTest.java
@@ -1,6 +1,7 @@
 package net.gazeplay;
 
 import javafx.scene.image.Image;
+import net.gazeplay.commons.utils.ApplicationIconImageLocator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.testfx.framework.junit5.ApplicationExtension;
